tests/test_api.py
"""
Basic Tests for AI Chatbot
"""

import pytest
from fastapi.testclient import TestClient


def test_placeholder():
    """Placeholder test - add more tests as needed"""
    assert True


# Add tests when running locally:
# 
# from app.main import app
# client = TestClient(app)
# 
# def test_health():
#     response = client.get("/api/v1/health")
#     assert response.status_code == 200
#     assert response.json()["status"] == "healthy"
# 
# def test_chat():
#     response = client.post("/api/v1/chat", json={"message": "Hello"})
#     assert response.status_code == 200
#     assert "response" in response.json()
