requirements.txt
# Core LangChain
langchain==0.3.14
langchain-openai==0.2.14
langchain-anthropic==0.3.0
langchain-community==0.3.14
langchain-core==0.3.28

# Vector Store
chromadb==0.5.23

# Embeddings
tiktoken==0.8.0

# Document Loaders
pypdf==5.1.0
python-docx==1.1.2

# API Framework
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# HTTP Client
httpx==0.28.1
